Signed-off-by: kunalrshah <kunal.rajnikant.shah@gmail.com>
Some minor changes to the CSS.
Ideally would like to change the helper.js as well - to utilize the grid layout styles
row, col-1 ... col-12 as appropriate.

While the project is submitted, reviewed and task completed - I intend to keep refining this to make a more interactive resume and also make it adjust itself as per the form-factor on which it is viewed.